refactor

+++ internal/handlers/requests/productBatchRequest.go
@@ -7,7 +7,7 @@ type ProductBatchRequest struct {
 	DueDate            string  `json:"due_date"`
 	InitialQuantity    int     `json:"initial_quantity"`
 	ManufacturingDate  string  `json:"manufacturing_date"`
-	ManufacturingHour  float64 `json:"manufacturing_hour"`
+	ManufacturingHour  int     `json:"manufacturing_hour"`
 	MinimumTemperature float64 `json:"minimum_temperature"`
 	ProductID          int     `json:"product_id"`
 	SectionID          int     `json:"section_id"`
